project1/requirements.txt
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
mysql-connector-python==8.0.25
protobuf==5.29.3
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.9.1
python-dotenv==1.0.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
